requirements.txt
Django
djangorestframework
django-parler
psycopg[binary]
python-dotenv
setuptools
unidecode
